requirements.txt
# Latest version with ansible < 4 support
ansible-modules-hashivault==4.6.3
